requirements.txt
msal
requests
pandas
openai
python-dotenv
